fix(searchEnabled): modify searchEnabled watch to evaluate attrs rather than scope property
searchEnabled was not reacting to changes in it's value.  This is because searchEnabled has a scope.$watch on it but searchEnabled was set as a property on $select not the scope and the watch would never fire.  Now the watch evaluates the attrs.searchEnabled for changes and properly reacts to changes in it's value.

Closes #505